Keep robber_hex a plain int, as its annotation says
move_robber still wrapped the value in np.uint16, so BoardState.robber_hex was
an int at setup and a numpy scalar after the first robber move. Missed by the
perf pass, which rewrote np.uint8/np.int8 but not np.uint16.

Harmless arithmetically -- everything reads it through int() -- but the field is
annotated int, and the verifier's snapshot showed the type changing mid-game.

Behaviour unchanged: tools/regression.py digests are identical.

+++ catan/interface.py
@@ -292,7 +292,7 @@ def move_robber(gs: GameState, rng: np.random.Generator, hex_id: int, victim_pid
             gs.players[gs.current_player_idx].hand[random_card] += 1
             gs.players[victim_pid].hand[random_card] -= 1
 
-    gs.board.robber_hex = np.uint16(hex_id)
+    gs.board.robber_hex = hex_id
